chore: Update GameCard.jsx to improve game platform footer styling
This commit updates the GameCard.jsx file to improve the styling of the game platform footer. The flexWrap property is added to ensure that the platform icons wrap to the next line if they exceed the container width. Additionally, the textAlign property is set to center to align the platform icons in the middle of the footer. This change enhances the visual presentation of the game platform information in the GameCard component.